version5

+++ assign4.pde
@@ -124,7 +124,7 @@ void keyPressed() {
 }
 
 /*---------Make Alien Function-------------*/
-void alienMaker(int alienNum, float inputCol) {
+void alienMaker(int alienNum, int inputCol) {
   for (int i =0; i<this.alienNum; i++) {
     int col = int(i % inputCol;)
     int row = int(i / inputCol;)
